refactor(rootage): rename dialog to alert-dialog, content-dialog to dialog
Rename component spec ids to match the Figma naming: the alert-style
dialog becomes alert-dialog, and the general-purpose content-dialog
reclaims the dialog id (content-dialog-close-button follows as
dialog-close-button). Regenerate component vars across css,
lynx-css, qvism-preset, and lynx-qvism-preset, remove stale
generated files, and update the two vars imports in qvism recipes.
Recipe names and CSS output are unchanged.
