+++ packages/css/vars/component/dialog-close-button.mjs
