+++ packages/lynx-css/vars/component/dialog-close-button.d.ts
