+++ packages/lynx-css/vars/component/dialog-close-button.mjs
