+++ packages/lynx-css/vars/component/index.d.ts
@@ -5,6 +5,7 @@ export { vars as actionChip } from "./action-chip";
 export { vars as actionSheetCloseButton } from "./action-sheet-close-button";
 export { vars as actionSheetItem } from "./action-sheet-item";
 export { vars as actionSheet } from "./action-sheet";
+export { vars as alertDialog } from "./alert-dialog";
 export { vars as attachmentInputDropzone } from "./attachment-input-dropzone";
 export { vars as attachmentInputItemActionButton } from "./attachment-input-item-action-button";
 export { vars as attachmentInputItemRemoveButton } from "./attachment-input-item-remove-button";
@@ -23,11 +24,10 @@ export { vars as checkbox } from "./checkbox";
 export { vars as checkmark } from "./checkmark";
 export { vars as chipTablist } from "./chip-tablist";
 export { vars as chip } from "./chip";
-export { vars as contentDialogCloseButton } from "./content-dialog-close-button";
-export { vars as contentDialog } from "./content-dialog";
 export { vars as contentPlaceholder } from "./content-placeholder";
 export { vars as contextualFloatingButton } from "./contextual-floating-button";
 export { vars as controlChip } from "./control-chip";
+export { vars as dialogCloseButton } from "./dialog-close-button";
 export { vars as dialog } from "./dialog";
 export { vars as divider } from "./divider";
 export { vars as extendedActionSheetCloseButton } from "./extended-action-sheet-close-button";
