+++ packages/lynx-qvism-preset/src/vars/component/index.mjs
@@ -5,6 +5,7 @@ export { vars as actionChip } from "./action-chip.mjs";
 export { vars as actionSheetCloseButton } from "./action-sheet-close-button.mjs";
 export { vars as actionSheetItem } from "./action-sheet-item.mjs";
 export { vars as actionSheet } from "./action-sheet.mjs";
+export { vars as alertDialog } from "./alert-dialog.mjs";
 export { vars as attachmentInputDropzone } from "./attachment-input-dropzone.mjs";
 export { vars as attachmentInputItemActionButton } from "./attachment-input-item-action-button.mjs";
 export { vars as attachmentInputItemRemoveButton } from "./attachment-input-item-remove-button.mjs";
@@ -23,11 +24,10 @@ export { vars as checkbox } from "./checkbox.mjs";
 export { vars as checkmark } from "./checkmark.mjs";
 export { vars as chipTablist } from "./chip-tablist.mjs";
 export { vars as chip } from "./chip.mjs";
-export { vars as contentDialogCloseButton } from "./content-dialog-close-button.mjs";
-export { vars as contentDialog } from "./content-dialog.mjs";
 export { vars as contentPlaceholder } from "./content-placeholder.mjs";
 export { vars as contextualFloatingButton } from "./contextual-floating-button.mjs";
 export { vars as controlChip } from "./control-chip.mjs";
+export { vars as dialogCloseButton } from "./dialog-close-button.mjs";
 export { vars as dialog } from "./dialog.mjs";
 export { vars as divider } from "./divider.mjs";
 export { vars as extendedActionSheetCloseButton } from "./extended-action-sheet-close-button.mjs";
