+++ packages/qvism-preset/src/vars/component/alert-dialog.d.ts
@@ -1,4 +1,4 @@
-export const vars = {
+export declare const vars: {
   "base": {
     "enabled": {
       "backdrop": {
@@ -13,9 +13,9 @@ export const vars = {
       "content": {
         "color": "var(--seed-color-bg-layer-floating)",
         "cornerRadius": "var(--seed-radius-r5)",
-        "widthFraction": "0.9",
-        "maxHeightFraction": "0.8",
-        "marginX": "var(--seed-dimension-x5)",
+        "marginX": "var(--seed-dimension-x8)",
+        "marginY": "var(--seed-dimension-x16)",
+        "maxWidth": "272px",
         "enterDuration": "var(--seed-duration-d4)",
         "enterTimingFunction": "var(--seed-timing-function-enter-expressive)",
         "enterOpacity": "0",
@@ -26,58 +26,26 @@ export const vars = {
       },
       "header": {
         "gap": "var(--seed-dimension-x1_5)",
-        "paddingX": "var(--seed-dimension-x6)",
-        "paddingTop": "var(--seed-dimension-x6)",
-        "paddingBottom": "var(--seed-dimension-x4)",
-        "closeButtonGap": "var(--seed-dimension-x1_5)"
-      },
-      "body": {
-        "paddingX": "var(--seed-dimension-x6)",
-        "paddingBottom": "var(--seed-dimension-x12)",
-        "strokeDuration": "var(--seed-duration-color-transition)",
-        "strokeTimingFunction": "var(--seed-timing-function-easing)"
+        "paddingX": "var(--seed-dimension-x5)",
+        "paddingTop": "var(--seed-dimension-x5)"
       },
       "footer": {
         "gap": "var(--seed-dimension-x2)",
-        "paddingX": "var(--seed-dimension-x6)",
+        "paddingX": "var(--seed-dimension-x5)",
         "paddingTop": "var(--seed-dimension-x4)",
-        "paddingBottom": "var(--seed-dimension-x6)"
+        "paddingBottom": "var(--seed-dimension-x5)"
       },
       "title": {
         "color": "var(--seed-color-fg-neutral)",
-        "fontSize": "var(--seed-font-size-t8)",
-        "lineHeight": "var(--seed-line-height-t8)",
+        "fontSize": "var(--seed-font-size-t7)",
+        "lineHeight": "var(--seed-line-height-t7)",
         "fontWeight": "var(--seed-font-weight-bold)"
       },
       "description": {
-        "color": "var(--seed-color-fg-neutral-muted)",
+        "color": "var(--seed-color-fg-neutral)",
         "fontSize": "var(--seed-font-size-t5)",
         "lineHeight": "var(--seed-line-height-t5)",
         "fontWeight": "var(--seed-font-weight-regular)"
-      },
-      "closeButton": {
-        "fromTop": "var(--seed-dimension-x7)",
-        "fromRight": "var(--seed-dimension-x6)"
-      }
-    },
-    "scrolled": {
-      "body": {
-        "strokeColor": "var(--seed-color-stroke-neutral-muted)",
-        "strokeWidth": "1px"
-      }
-    }
-  },
-  "sizeMedium": {
-    "enabled": {
-      "content": {
-        "maxWidth": "480px"
-      }
-    }
-  },
-  "sizeLarge": {
-    "enabled": {
-      "content": {
-        "maxWidth": "800px"
       }
     }
   }
