+++ packages/qvism-preset/src/vars/component/dialog-close-button.d.ts
