+++ packages/qvism-preset/src/vars/component/dialog-close-button.mjs
